+++ index.html
@@ -232,7 +232,6 @@
     <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-    <script src="/js/main.js"></script>
 </body>
-
+<script src="./js/main.js" type="module"></script>
 </html>
