+++ js/data.js
@@ -1,0 +1,4 @@
+const MAX_COUNT_OF_USERS = 25;
+const DESCRIPTION_LIST = ['Милые котята', 'Собачки', 'Новости', 'Веселое видео', 'Новая игра', 'Мем'];
+
+export {MAX_COUNT_OF_USERS, DESCRIPTION_LIST};
